+++ app/lib/tool/test_profile/resolver.dart
@@ -31,8 +31,9 @@ Future<List<ResolvedVersion>> resolveVersions(
     await pubCacheDir.create();
 
     final toolEnv = await ToolEnvironment.create(
-      dartSdkDir: envConfig.previewDartSdkDir,
-      flutterSdkDir: envConfig.flutterSdkDir,
+      dartSdkDir: envConfig.previewDartSdkDir ?? envConfig.stableDartSdkDir,
+      flutterSdkDir:
+          envConfig.previewFlutterSdkDir ?? envConfig.stableFlutterSdkDir,
       pubCacheDir: pubCacheDir.path,
     );